+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =====================
 
+v1.2
+---------------------
+* NEW: Bootstrap4 Themes support.
+* NEW: RouteAccessControl filter autoCreatePermission option in debug mode.
+
 v1.1.3
 ---------------------
 * Bugfix: Compatibility with Adminlte v2.6.0 (conflict of $.fn.tree plugin name). @ap